+++ HW7-2/app/src/main/java/com/lawk/GLocation/Camera.java
@@ -1,0 +1,25 @@
+package com.lawk.GLocation;
+
+public class Camera {
+    private String cameraName;
+    private String imageURL;
+    private String type;
+
+    public Camera(String cameraName, String imageURL, String type) {
+        this.cameraName = cameraName;
+        this.imageURL = imageURL;
+        this.type = type;
+    }
+
+    public String getCameraName() {
+        return cameraName;
+    }
+    public String getImageURL() {
+        return imageURL;
+    }
+
+
+    public String getType() {
+        return type;
+    }
+}
