+++ HW7-2/app/src/main/java/com/lawk/GLocation/CameraCoordinates.java
@@ -1,0 +1,21 @@
+package com.lawk.GLocation;
+
+public class CameraCoordinates {
+    private double x;
+    private double y;
+
+    public CameraCoordinates(double x, double y) {
+        this.x = x;
+        this.y = y;
+    }
+
+    public double getX() {
+        return x;
+    }
+
+    public double getY() {
+        return y;
+    }
+
+
+}
