+++ HW7-2/app/src/main/res/layout/activity_main.xml
