+++ HW7-2/app/src/main/res/layout/cardview.xml
