+++ HW7-2/app/src/main/res/mipmap-mdpi/ic_launcher.png
