+++ HW7-2/app/src/main/res/mipmap-xxxhdpi/ic_launcher.png
