+++ HW7-2/app/src/main/res/values/colors.xml
