+++ HW7-2/app/src/main/res/values/google_maps_api+key.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="google_maps_key" templateMergeStrategy="preserve" translatable="false">AIzaSyDu1DWiervfNDhA9grVW8-RIDSpQZhW998</string>
+</resources>
