+++ HW7-2/app/src/main/res/values/strings.xml
@@ -1,0 +1,3 @@
+<resources>
+    <string name="app_name">GLocation</string>
+</resources>
