+++ HW7-2/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Fri Jun 07 05:12:48 PDT 2019
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-4.10.1-all.zip
