--- HW7/app/src/main/java/com/lawk/location/CameraCoordinates.java
@@ -1,21 +1,0 @@
-package com.lawk.location;
-
-public class CameraCoordinates {
-    private double x;
-    private double y;
-
-    public CameraCoordinates(double x, double y) {
-        this.x = x;
-        this.y = y;
-    }
-
-    public double getX() {
-        return x;
-    }
-
-    public double getY() {
-        return y;
-    }
-
-
-}
