--- HW7/app/src/main/res/values/strings.xml
@@ -1,6 +1,0 @@
-<resources>
-    <string name="app_name">Google</string>
-
-    <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
-</resources>
